requirements.txt
pillow
openai
python-dotenv
flickrapi
tqdm
tiktoken
huggingface_hub
